add missing translation to berlimap module

+++ db_update.php
@@ -138,36 +138,50 @@ echo 'Creating ws_fieldtype entry for new uitype...';
 $adb->query("INSERT INTO vtiger_ws_fieldtype (`uitype` ,`fieldtype`) VALUES ('cr16', 'autocompletedtext')");
 echo 'done<br>';
 
-echo "<br>create table for new Mailchimp version ";
-$query = "CREATE TABLE IF NOT EXISTS `vtiger_mailchimp_synced_entities` (
-  `crmid` int(11) NOT NULL,
-  `mcgroupid` int(11) NOT NULL,
-  `recordid` int(11) NOT NULL,
-  KEY `recordidx` (`recordid`),
-  KEY `crmid` (`crmid`),
-  CONSTRAINT `fk_1_vtiger_mailchimp_synced_entities` FOREIGN KEY (`crmid`) REFERENCES `vtiger_crmentity` (`crmid`) ON DELETE CASCADE
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci";
-$adb->pquery($query, array());
-echo "create table done.<br>";
+$moduleInstance = Vtiger_Module::getInstance("Mailchimp");
+if($moduleInstance) {
+    echo "<br>Updating Mailchimp module ";
+    updateVtlibModule("Mailchimp", "packages/vtiger/mandatory/Mailchimp.zip");
+    echo "- done";
+    
+    echo "<br>create table for new Mailchimp version ";
+    $query = "CREATE TABLE IF NOT EXISTS `vtiger_mailchimp_synced_entities` (
+      `crmid` int(11) NOT NULL,
+      `mcgroupid` int(11) NOT NULL,
+      `recordid` int(11) NOT NULL,
+      KEY `recordidx` (`recordid`),
+      KEY `crmid` (`crmid`),
+      CONSTRAINT `fk_1_vtiger_mailchimp_synced_entities` FOREIGN KEY (`crmid`) REFERENCES `vtiger_crmentity` (`crmid`) ON DELETE CASCADE
+    ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci";
+    $adb->pquery($query, array());
+    echo "create table done.<br>";
 
-echo "<br>delete not needed files for new Mailchimp version ";
-require_once('config.inc.php');
-$filePath = "modules/Mailchimp/actions/logfileWriter.php";
-$file = $root_directory.''.$filePath;
-unlink($file);
-$filePath = "modules/Mailchimp/actions/MailchimpSyncStep1.php";
-$file = $root_directory.''.$filePath;
-unlink($file);
-$filePath = "modules/Mailchimp/actions/MailchimpSyncStep2.php";
-$file = $root_directory.''.$filePath;
-unlink($file);
-$filePath = "modules/Mailchimp/actions/MailchimpSyncStep3.php";
-$file = $root_directory.''.$filePath;
-unlink($file);
-$filePath = "modules/Mailchimp/actions/MailchimpSyncStep4.php";
-$file = $root_directory.''.$filePath;
-unlink($file);
-echo "file deletion done.<br>";
+    echo "<br>delete not needed files for new Mailchimp version ";
+    require_once('config.inc.php');
+    $filePath = "modules/Mailchimp/actions/logfileWriter.php";
+    $file = $root_directory.''.$filePath;
+    unlink($file);
+    $filePath = "modules/Mailchimp/actions/MailchimpSyncStep1.php";
+    $file = $root_directory.''.$filePath;
+    unlink($file);
+    $filePath = "modules/Mailchimp/actions/MailchimpSyncStep2.php";
+    $file = $root_directory.''.$filePath;
+    unlink($file);
+    $filePath = "modules/Mailchimp/actions/MailchimpSyncStep3.php";
+    $file = $root_directory.''.$filePath;
+    unlink($file);
+    $filePath = "modules/Mailchimp/actions/MailchimpSyncStep4.php";
+    $file = $root_directory.''.$filePath;
+    unlink($file);
+    echo "file deletion done.<br>";
+}
+
+$moduleInstance = Vtiger_Module::getInstance("berlimap");
+if($moduleInstance) {
+    echo "<br>Updating berlimap module ";
+    updateVtlibModule("berlimap", "packages/vtiger/optional/berlimap.zip");
+    echo "- done";
+}
 
 echo "<br>update Tag version to 13.. ";
 $query = "UPDATE `vtiger_version` SET `tag_version` = 'berlicrm-1.0.0.13'";
